Fixed port detection

+++ src/main/java/com/gitlab/zachdeibert/jnet/NetworkClient.java
@@ -183,7 +183,7 @@ public final class NetworkClient extends AsyncDeserializer implements
      *             An I/O error has occurred
      */
     @Override
-    public short getPort() throws IOException
+    public int getPort() throws IOException
     {
         final SocketAddress server = socket.getRemoteSocketAddress();
         if (server == null)
@@ -192,7 +192,7 @@ public final class NetworkClient extends AsyncDeserializer implements
         }
         if (server instanceof InetSocketAddress)
         {
-            return (short) ((InetSocketAddress) server).getPort();
+            return ((InetSocketAddress) server).getPort();
         }
         else
         {
